\name{banknote}
\alias{banknote}
\docType{data}

\title{Swiss banknotes data}

\description{
The data set contains six measurements made on 100 genuine and 100 counterfeit old-Swiss 1000-franc bank notes.}

\usage{data(banknote)}

\format{A data frame with the following variables:

\describe{
\item{Status}{the status of the banknote: \code{genuine} or \code{counterfeit}}
\item{Length}{Length of bill (mm)}
\item{Left}{Width of left edge (mm)}
\item{Right}{Width of right edge (mm)}
\item{Bottom}{Bottom margin width (mm)}
\item{Top}{Top margin width (mm)}
\item{Diagonal}{Length of diagonal (mm)}
}

}

\source{
Flury, B. and Riedwyl, H. (1988). \emph{Multivariate Statistics: A practical approach.} London: Chapman & Hall, Tables 1.1 and 1.2, pp. 5-8
}
% \examples{}
\keyword{datasets}
