\name{chevron}
\alias{chevron}
\title{Simulated minefield data}
\usage{data(chevron)}
\description{A set of simulated bivariate minefield data
  (1104 observations).}
\references{
A. Dasgupta and A. E. Raftery (1998).
  Detecting features in spatial point processes with clutter via model-based
  clustering. 
 \emph{Journal of the American Statistical Association 93:294-302}. 

C. Fraley and A.E. Raftery (1998).
   \emph{Computer Journal 41:578-588}.

G. J. McLachlan and D. Peel (2000).
   \emph{Finite Mixture Models}, Wiley, pages 110-112.
}

\keyword{datasets}

