\name{print.emclust}
\alias{print.emclust1}
\alias{print.bic}

\title{Print methods for BIC values}
\description{
  Print objects from \code{emclust}, \code{emclust1} or \code{bic}.
}
\usage{
print.emclust(x, ...)
print.emclust1(x, ...)
print.bic(x, ...)
}

\arguments{
  \item{x}{Object containing BIC values}
  \item{\dots}{Further arguments to print.default}
}

\keyword{clustering}
