\name{vecnorm}
\alias{vecnorm}
\title{Norm of a vector}
\description{
  Function returns norm of a vector.
}
\usage{
vecnorm(x, p = 2)
}
\arguments{
  \item{x}{Vector.}
  \item{p}{For \code{p = 2} (default), the Euclidean norm is returned.}
}
\value{
  Scalar, the norm of the vector.
}
\keyword{cluster}
