\name{diabetes}
\alias{diabetes}
\docType{data}

\title{Diabetes data}

\description{The data set contains three measurements made on 145 non-obese adult patients classified into three groups.}

\usage{data(diabetes)}

\format{A data frame with the following variables:
\describe{
\item{class}{The type of diabete: \code{Normal}, \code{Overt}, and \code{Chemical}.}
\item{glucose}{Area under plasma glucose curve after a three hour oral glucose tolerance test (OGTT).}
\item{insulin}{Area under plasma insulin curve after a three hour oral glucose tolerance test (OGTT).}
\item{sspg}{Steady state plasma glucose.}
}
}

\details{This dataset is \emph{not correct} and it is provided here only for backward compatibility. Please refer to the correct version available in package \pkg{rrcov}.}

\source{Reaven, G. M. and Miller, R. G. (1979). An attempt to define the nature of chemical diabetes using a multidimensional analysis. \emph{Diabetologia} 16:17-24.}

\keyword{datasets}
