\name{diabetes}
\alias{diabetes.class}
\non_function{}
\title{Diabetes data}
\usage{data(diabetes)}
\description{
Diabetes data from Reaven and Miller. Number of objects: 145; 3
variables. Three classes.
}
\references{G.M. Reaven and R.G. Miller, \emph{Diabetologica} 16:17-24 (1979).
  }

\keyword{datasets}
