\name{EuroUnemployment}
\alias{EuroUnemployment}
\docType{data}

\title{Unemployment data for European countries in 2014}

\description{
The data set contains unemployment rates for 31 European countries for the year 2014.}

\usage{data(EuroUnemployment)}

\format{A data frame with the following variables:

\describe{

\item{TUR}{Total unemployment rate, i.e. percentage of unemployed persons aged 15-74 in the economically active population.}
\item{YUR}{Youth unemployment rate, i.e. percentage of unemployed persons aged 15-24 in the economically active population.}
\item{LUR}{Long-term unemployment rate, i.e. percentage of unemployed persons who have been unemployed for 12 months or more.}
}

}

\source{Dataset downloaded from EUROSTAT \url{https://ec.europa.eu/eurostat}.}

\keyword{datasets}
