\name{mclust-internal}
\title{Internal MCLUST functions}
\alias{pickBIC}
\alias{bicFill}
\alias{grid1}
\alias{grid2}
\alias{mvn2plot}
\alias{vecnorm}
\alias{traceW}
\alias{qclass}
\alias{unchol}
\alias{shapeO}
\alias{orth2}
\alias{charconv}
\alias{[.mclustBIC}
\alias{[.mclustDAtest}
\description{
  Internal functions.
}
\details{
  Not intended to be called directly by users.
}
\keyword{internal}
