\name{mclust-internal}
\title{Internal MCLUST functions}
\alias{mvn2plot}
\alias{vecnorm}
\alias{traceW}
\alias{unchol}
\alias{shapeO}
\alias{orth2}
\alias{mclustProjControl}
\alias{mclust2DplotControl}
\alias{[.EMclust}
\alias{[.EMclustN}
\alias{nextPerm}
\alias{charconv}
\alias{estep2}
\description{
  Internal tools functions.
}
\details{
  These are not to be called by the user directly.
}
\keyword{internal}
