\name{mclust-internal}

\title{Internal MCLUST functions}

\alias{pickBIC}
\alias{bicFill}
\alias{grid1}
\alias{grid2}
\alias{mvn2plot}
\alias{vecnorm}
\alias{traceW}
\alias{qclass}
\alias{unchol}
\alias{shapeO}
\alias{orth2}
\alias{charconv}
\alias{[.mclustBIC}
\alias{checkModelName}
\alias{balanced.folds}
\alias{permute.rows}
\alias{projpar.MclustDR}
\alias{projdir.MclustDR}
\alias{mvdnorm}
\alias{ellipse}
\alias{eigen.decomp}
\alias{dmvnorm}
\alias{getParameters.MclustDA}

\description{
  Internal functions not intended to be called directly by users.
}

\keyword{internal}
