\name{mclust-internal}

\title{Internal MCLUST functions}

\alias{pickBIC}
\alias{bicFill}
\alias{grid1}
\alias{grid2}
\alias{mvn2plot}
\alias{vecnorm}
\alias{traceW}
\alias{qclass}
\alias{unchol}
\alias{shapeO}
\alias{orth2}
\alias{charconv}
\alias{[.mclustBIC}
\alias{checkModelName}
\alias{balancedFolds}
\alias{permuteRows}
\alias{projpar.MclustDR}
\alias{projdir.MclustDR}
%\alias{mvdnorm}
\alias{ellipse}
\alias{eigen.decomp}
\alias{getParameters.MclustDA}
\alias{as.Mclust}
\alias{as.Mclust.default}
\alias{as.Mclust.densityMclust}
\alias{as.densityMclust}
\alias{as.densityMclust.default}
\alias{as.densityMclust.Mclust}
\alias{count}

\description{
  Internal functions not intended to be called directly by users.
}

\keyword{internal}
