\name{summary.Mclust}
\alias{summary.Mclust}
\title{Very brief summary of an Mclust object.}
\description{
  Function gives a brief summary of an Mclust object: the type of model
  that is picked and the number of clusters.
}
\usage{
summary.Mclust(object, ...)
}
\arguments{
  \item{object}{The result of a call to
    function \code{Mclust}.}
  \item{\dots}{Not used.}
}
\keyword{cluster}
