\name{wreath}
\alias{wreath}

\title{Data Simulated from a 14-Component Mixture}

\usage{data(wreath)}

\description{
A dataset consisting of 1000 observations drawn from a 14-component 
normal mixture in which the covariances of the components have the
same size and shape but differ in orientation.   
}
\references{
  C. Fraley,  A. E. Raftery and R. Wehrens (2005).
  Incremental model-based clustering for large datasets with small clusters.
  \emph{Journal of Computational and Graphical Statistics 14:1:18}.
}

\keyword{datasets}
